requirements.txt
Pillow
rblx_open_cloud
Requests